+++ pkg/db/dbconn/conn.go
@@ -1,0 +1,6 @@
+package dbconn
+
+type DbConn struct {
+	Type string
+	Conn interface{}
+}
