+++ pkg/db/tx_func/tx_func.go
@@ -1,0 +1,3 @@
+package tx_func
+
+type TxFunc func() error
